Remove registerServiceWorker (for now) and boilerplate
I decided to comment out the registerServiceWorker for now since it's new to createReactApp and I don't really know what it does.
I read into it briefly, and it sounds like it will create issues with users on iOS that are trying to use this react app.
I also removed other boilerplate code that I don't plan on using.